+++ parser_library/src/processing/statement_processors/lookahead_processor.cpp
@@ -235,7 +235,7 @@ void lookahead_processor::assign_data_def_attributes(context::id_index symbol_na
 
     context::symbol_attributes::type_attr type =
         ebcdic_encoding::a2e[(unsigned char)data_op->value->get_type_attribute()];
-    ;
+
     context::symbol_attributes::len_attr len = context::symbol_attributes::undef_length;
     context::symbol_attributes::scale_attr scale = context::symbol_attributes::undef_scale;